+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Improved
+- **deps**: Upgrade @types/node to 22.13.11
+- **deps**: Upgrade @valkyriestudios/utils to 12.35.0
+- **deps**: Upgrade eslint to 9.23.0
+- **deps**: Upgrade typescript-eslint to 8.27.0
+
 ## [10.1.0] - 2025-03-09
 ### Added
 - **feat**: Added support for using schemas in conditional blocks, eg: